requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
pycaret>=3.3.0
mlflow>=2.12
flask>=3.0
python-dotenv>=1.0
scipy>=1.10
flasgger==0.9.7.1

